Development

+++ project/core/src/com/tdt4240/RawHeroes/independent/inputListeners/TouchDown.java
@@ -1,5 +1,5 @@
 package com.tdt4240.RawHeroes.independent.inputListeners;
 
 public interface TouchDown {
-    public boolean touchDown(int screenX, int screenY, int pointer, int button);
+    public boolean touchDown(float screenX, float screenY, int pointer, int button);
 }
